split Display impls into src/display.rs; removed useless FnArgsIns struct
